moving duplicated code into an extra-function
